Move factory method to HitboxSceneNode
